REFACTOR: move Okada sources into static directory
